+++ src/lib/parser/tokenize.ts
@@ -17,6 +17,7 @@ export function* tokenize(stream: string): Generator<TokenizeResult> {
 
     let value;
     let buffer: string = '';
+
     // let input: string = '';
 
     function consumeWhiteSpace(): number {
@@ -319,8 +320,11 @@ export function* tokenize(stream: string): Generator<TokenizeResult> {
                 // EOF
                 if (!(value = next())) {
                     // end of stream ignore \\
-                    yield pushToken(buffer);
-                    buffer = '';
+                    if (buffer.length > 0) {
+
+                        yield pushToken(buffer);
+                        buffer = '';
+                    }
                     break;
                 }
 
@@ -330,37 +334,70 @@ export function* tokenize(stream: string): Generator<TokenizeResult> {
             case "'":
                 yield* consumeString(value);
                 break;
+            case '^':
             case '~':
             case '|':
+            case '$':
+
+                if (value == '|' && peek() == '|') {
+
+                    next();
+                    yield pushToken('', EnumToken.ColumnCombinatorTokenType);
+                    buffer = '';
+                    break;
+                }
 
                 if (buffer.length > 0) {
                     yield pushToken(buffer);
                     buffer = '';
                 }
+
                 buffer += value;
 
-                if (!(value = next())) {
+                if (!(value = peek())) {
                     yield pushToken(buffer);
                     buffer = '';
                     break;
                 }
-                if (value == '=') {
 
-                    buffer += value;
-                    yield pushToken(buffer, buffer[0] == '~' ? EnumToken.IncludesTokenType : EnumToken.DashMatchTokenType);
+                // ~=
+                // ^=
+                // $=
+                // |=
+                if (peek() == '=') {
+
+                    next();
+                    switch (buffer.charAt(0)) {
+
+                        case '~':
+                            yield pushToken(buffer, EnumToken.IncludeMatchTokenType);
+                            break;
+                        case '^':
+                            yield pushToken(buffer, EnumToken.StartMatchTokenType);
+                            break;
+                        case '$':
+                            yield pushToken(buffer, EnumToken.EndMatchTokenType);
+                            break;
+                        case '|':
+                            yield pushToken(buffer, EnumToken.DashMatchTokenType);
+                            break;
+                    }
 
                     buffer = '';
                     break;
                 }
 
                 yield pushToken(buffer);
+                buffer = '';
 
-                while (isWhiteSpace(value.charCodeAt(0))) {
+                // yield pushToken(buffer);
+                //
+                // while (isWhiteSpace(value.charCodeAt(0))) {
+                //
+                //     value = next();
+                // }
 
-                    value = next();
-                }
-
-                buffer = value;
+                // buffer = value;
                 break;
 
             case '>':
@@ -405,7 +442,17 @@ export function* tokenize(stream: string): Generator<TokenizeResult> {
                     buffer = '';
                 }
 
-                if (value == ':' && ':' == peek()) {
+                const val: string = peek();
+
+                if (val == '=') {
+
+                    next();
+
+                    yield pushToken(value + val, EnumToken.ContainMatchTokenType);
+                    break;
+                }
+
+                if (value == ':' && ':' == val) {
 
                     buffer += value + next();
                     break;
@@ -414,7 +461,7 @@ export function* tokenize(stream: string): Generator<TokenizeResult> {
                 yield pushToken(value);
                 buffer = '';
 
-                if (['+', '*', '/'].includes(value)   && isWhiteSpace(peek().charCodeAt(0))) {
+                if (['+', '*', '/'].includes(value) && isWhiteSpace(peek().charCodeAt(0))) {
 
                     yield pushToken(next());
                 }